Rename Python HNSW searcher module to match class naming convention
