+++ iris-mpc-py/src/py_hnsw/pyclasses/mod.rs
@@ -1,5 +1,5 @@
 pub mod graph_store;
-pub mod hawk_searcher;
+pub mod hnsw_searcher;
 pub mod iris_code;
 pub mod iris_code_array;
 pub mod plaintext_store;
